Use some HTML5 localstorage and JS to ensure that posts never get lost again.
Upon editing forms in the admin, JS will start auto saving your form data every second. If you come back to the same page (based on window.location.href) and there is an entry in local storage, you will be given an opportunity to recover or discard the stored info.